modif

+++ lib/my/my_strncat.c
@@ -9,17 +9,17 @@
 #include <unistd.h>
 #include "my.h"
 
-char *concat_n(char *s1, char *s2, int n)
+char *concat_n(char *s1, char *s2, size_t n)
 {
-	int c = 0;
+	size_t c = 0;
 	char *str = malloc(sizeof(char) * (my_strlen(s1) + n + 1));
 
 	if (str == NULL)
 		return (NULL);
-	for (int i = 0; s1[i] != '\0'; i++) {
+	for (size_t i = 0; s1[i] != '\0'; i++) {
 		str[c] = s1[i];
 		c++;
-	} for (int j = 0; j != n; j++) {
+	} for (size_t j = 0; j != n; j++) {
 		str[c] = s2[j];
 		c++;
 	}
